Protect APIs

+++ src/controller/cronController.js
@@ -5,13 +5,15 @@ const mongoose = require('mongoose');
 module.exports = {
   async executeClearGamificationDrop(req, res){
 
-    try{
-      const mongo = await mongoose.createConnection(process.env.MONGODB_URI);
-      const dropCollection = await mongo.dropCollection('gamifications');
-
-      return res.send(`Dropped - ${dropCollection}`);
-    }catch{
-      return res.send(`error`);
+    if(process.env.CRON_TOKEN === process.env.CRON_TOKEN_ENV){
+      try{
+        const mongo = await mongoose.createConnection(process.env.MONGODB_URI);
+        const dropCollection = await mongo.dropCollection('gamifications');
+  
+        return res.send(`Dropped - ${dropCollection}`);
+      }catch{
+        return res.send(`error`);
+      }
     }
     
     // const gamificationCollections = mongoose.connection.collections["gamifications"];
